lib: cockpit: split location into a separate class
This converts cockpit.location to a proper TypeScriptified class, not
intended yet for usage as an ESM module as for that the Location class
needs to be improved and modernized.
